implement review - separate listing logic and add unit tests
